requirements.txt
aiosmtplib==2.0.1
alembic==1.9.1
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==4.2.4
certifi==2023.5.7
click==8.1.3
cloudinary==1.33.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.99.1
fastapi-mail==1.3.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mirakuru==2.5.1
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
port-for==0.7.0
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
asyncpg==0.29.0
pyasn1==0.5.0
pydantic==1.10.4
pytest==7.2.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1
pytest-postgresql==5.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-slugify==8.0.1
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tenacity==8.2.2
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
python-multipart==0.0.6

# WebSocket and Redis support for real-time bidding
redis==4.6.0
websockets==11.0.3
celery==5.3.0

# Phase 6: Mobile and notification features
qrcode==7.4.2
Pillow==10.0.0

# SMS notifications
twilio==8.2.0

# Phase 5: Price History and Alerts
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.5
flower==2.0.1

# Phase 6: ML-Powered Recommendations
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
prophet==1.1.4
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
